solutions/aoc2021/src/test/java/net/anomalyxii/aoc/aoc2021/Day17Test.java
package net.anomalyxii.aoc.aoc2021;

import net.anomalyxii.aoc.context.SolutionContext;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class Day17Test {

    private static final SolutionContext EXAMPLE = SolutionContext.example(2021, 17);
    private static final SolutionContext LIVE = SolutionContext.live(2021, 17);

    // ****************************************
    // Test Methods
    // ****************************************

    // calculateAnswerForPart1

    @Test
    void calculateAnswerForPart1_Example() {
        // arrange
        final Day17 challenge = new Day17();

        // act
        final long answer = challenge.calculateAnswerForPart1(EXAMPLE);

        // assert
        assertThat(answer)
                .isEqualTo(45L);
    }

    @Test
    void calculateAnswerForPart1_Live() {
        // arrange
        final Day17 challenge = new Day17();

        // act
        final long answer = challenge.calculateAnswerForPart1(LIVE);

        // assert
        assertThat(answer)
                .isEqualTo(6555L);
    }

    // calculateAnswerForPart2

    @Test
    void calculateAnswerForPart2_Example() {
        // arrange
        final Day17 challenge = new Day17();

        // act
        final long answer = challenge.calculateAnswerForPart2(EXAMPLE);

        // assert
        assertThat(answer)
                .isEqualTo(112L);
    }

    @Test
    void calculateAnswerForPart2_Live() {
        // arrange
        final Day17 challenge = new Day17();

        // act
        final long answer = challenge.calculateAnswerForPart2(LIVE);

        // assert
        assertThat(answer)
                .isEqualTo(4973L);
    }

}
